requirements.txt
albumentations==0.5.1
antlr4-python3-runtime==4.8
certifi==2020.11.8
cycler==0.10.0
dataclasses==0.6
decorator==4.4.2
efficientnet-pytorch==0.6.3
future==0.18.2
hydra-core==1.0.4
imageio==2.9.0
imgaug==0.4.0
importlib-resources==3.3.0
install==1.3.4
kiwisolver==1.3.1
matplotlib==3.3.3
munch==2.5.0
networkx==2.5
numpy==1.19.4
omegaconf==2.0.5
opencv-python-headless==4.4.0.46
packaging==20.7
Pillow==8.0.1
pretrainedmodels==0.7.4
pyparsing==2.4.7
python-dateutil==2.8.1
PyWavelets==1.1.1
PyYAML==5.3.1
scikit-image==0.17.2
scipy==1.5.4
segmentation-models-pytorch==0.1.2
Shapely==1.7.1
six==1.15.0
tifffile==2020.11.18
timm==0.1.20
torch-lr-finder==0.2.1
tqdm==4.54.1
typing-extensions==3.7.4.3
